barebox: add TARGETCC to Makefile
Barebox has some userspace tools located in the scripts folder. They
have to be compiled with the toolchain settings from the build system.
There was no machanism before to achieve this. We added the TARGETCC
variable to be set by the build system.

This should be improved to pass the hole userspace toolchain setup to
the target tools.

Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>