sdcard_image-phy.bbclass: add e2fsprogs dependency explicitly
This is done because the SDIMG_ROOTFS_TYPE mechanism does not pull the
e2fsprogs-native dependency if its configured as ext3. This patch can be
removed if the mechanism works completely. We dont want the user to add
ext3 image dependency in the local.conf. It should be enough to set
SDIMG_ROOTFS_TYPE.

Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>