rename: recipes-images to recipes-openembedded
Those recipes will be moved to meta-yogurt, our distribution layer. For
the time being we indicate the recipes dependency on openembedded as
follows. There is no clean way other than making the hole layer depend
on openembedded, what is not wanted in any way. Please accept this as an
intermediate solution.

Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>