conf: machine: Fix require in all machine files
If you use a machine file as a template and copy it to another layer, bitbake
won't find the file "include/phyimx6.inc" referenced by the require directive
in the machine configuration, because the file isn't located in the current
directory anymore. So after copying the file, you have to fix the path
yourself.

To avoid this failure in the first place, we use the full path in your own
machines files, too. E.g.:
    require conf/machine/include/phyimx6.inc

Signed-off-by: Stefan Christ <s.christ@phytec.de>