classes: change default filename of barebox.bin
The meta-phyam335x layer creates a link barebox.bin in the deploy directory.
To keep the deploy structure and filenames of the meta-phyimx6 layer in sync
with the meta-phyam335x layer and make development nicer, we change the default
barebox image name in the sdcard creation class and remove the extra symlink
name in the barebox recipe.

Signed-off-by: Stefan Christ <s.christ@phytec.de>