recipes-kernel: mx6: remove kernel module evbug
Remove the kernel module evbug from the imx6 defconfig, because it is normally
not needed and loaded automatically. The module writes all events on the input
devices in /dev/input/* to the kernel ring buffer. It fills rapidly with
messages like

[ 1401.455656] evbug: Event. Dev: input0, Type: 3, Code: 47, Value: 1
[ 1401.461840] evbug: Event. Dev: input0, Type: 3, Code: 54, Value: 482
[ 1401.468197] evbug: Event. Dev: input0, Type: 3, Code: 0, Value: 720

Apart from being annoying it's also a security risk, because passwords may end
up there.

Signed-off-by: Stefan Christ <s.christ@phytec.de>