classes: cleanup filesystem format in image class
The preferred filesystem, which is written onto the sdcard, should be set in
the image creation class and should be not part of the machine configuration.
In the meta-phyam335x layer the image creation class handles this the same way.

It also allows an easy switch to the ext4 filesystem format.