recipes-bsp: barebox: add far.env.bin script
Instead of maintaining the script reboot.env.bin (new name far.env.bin) in the
meta-phyam335x and in the meta-phyimx6 layer move it to the common barebox
recipe. The only drawback is that the script is always included in the barebox
environment whether or not you enable the meta-phyintern layers.

The script depends on the option 'debug-tweaks' in the variable
EXTRA_IMAGE_FEATURES.  It isn't included in a production build.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>