recipes: fix COMPATIBLE_MACHINE in barebox and linux-mainline
Remove the machine override suffix from variable COMPATIBLE_MACHINE in the
barebox_git.bb and linux-mainline.bb recipe, because this form hasn't an
effect.  The COMPATIBLE_MACHINE variable would be only applied, if the machine
is already set to 'mx6'. If MACHINE doesn't contain mx6, the COMPATIBLE_MACHINE
variable wouldn't be applied and the recipe is compatible with _all_ machines.
This is obviously wrong. In the last case the recipe shouldn't be considered,
because it's mx6 specific.

Furthermore this change allows to enable the meta-phyimx6 and meta-phyam335x
layer simultaneously in the bblayers.conf, because the recipe selection depends
only on the MACHINE variable which is defined in the local.conf.

Signed-off-by: Stefan Christ <s.christ@phytec.de>