recipes-bsp: barebox: fix bootsource in environment
The phyBOARD-MIRA has a the file init/bootsource in his default environment. In
this file the variable boot.default is initialized and the name 'mmc' is used
for the name of the sdcard boot configuration. Our sdcard boot configuration
file is named 'sd-ext3'. Therefore this patch replaces the file init/bootsource
with an empty dummy and the correct name of the boot configuration is already
set in the environment file 'config'.

Signed-off-by: Stefan Christ <s.christ@phytec.de>