sdcard_image-phy.bbclass: fix dependency
The deploy directory gets used for partitioned image creation. Yocto is
not quite design to work like that. We will get complicated dependency
problems in future if we try to integrate more Platforms in this
workflow. This patch repairs one of those depencies when recreating
images. In future we should switch to wic for partitioned images.

Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>