meta-phyintern: recipes: remove wildcard in bbappends
Using the wildcard '%' without a prefix in a bbappend filename is very
dangerous, because the bbappend is applied to all versions of the recipe in all
included layers. This will mostly break some existing recipes, which are not
considered by the bbappend author. E.g. the current barebox_%.bbappend in
meta-phyintern, which was written only for 'barebox_git.bb', was also applied
to the bb-file 'barebox_2014.07.0.bb' in the layer meta-phytec.

It's better to be very narrow with any bbappend wildcard-matching.

Signed-off-by: Stefan Christ <s.christ@phytec.de>