classes: add bbclass phygittag
The bbclass phygittag is a generic approach to fullfill our development needs.
It allows a developer to overwrite the package's git url and to use ${AUTOREV}
in his local.conf without interfering with the yocto/bitbake design principles.
He/She can

  - use the head of the branch instead of a tag version,
  - use another branch in the repository,
  - use another remote git repository or
  - use a local clone of the repository.

All possibilites with examples are documented in the appropriate
documentations.

An example, how a package maintainer has to use the bbclass, can be found in
the bbclass's comments.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>