recipes: change filename schema and use bbclass phygittag
Now we use the version of the package in the recipe's filename. This is useful
when we have to support multiple versions of a package for different machines,
because machine overrides lead to a confusing and complex variable hierarchy.
The machine selection is done with the variable COMPATIBLE_MACHINE in which all
supported machines are listed.

Furthermore the new bbclass phygittag is used to allow a developer to overwrite
the git url and SRCREV in his local.conf without setting the PV in the
local.conf, too, because setting PV the local.conf breaks the package selection
of bitbake.

Signed-off-by: Stefan Christ <s.christ@phytec.de>