barebox, barebox-ipl: use new filename schema
Now we use the version of the package in the recipe's filename. This is
useful when we have to support multiple versions of a package for
different machines, because machine overrides lead to a confusing and
complex variable hierarchy. The machine selection is done with the
variable COMPATIBLE_MACHINE in which all supported machines are listed.

Furthermore the new bbclass phygittag is used to allow a developer to
overwrite the git url and SRCREV in his local.conf without setting the
PV in the local.conf, too, because setting PV the local.conf breaks the
package selection of bitbake.

For the recipe we are not allowed to set AUTOREV, because this breaks
the possibility to work offline. AUTOREV should be set only in the
local.conf.

Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>