conf: machine: clarify phyflex-imx6-? machines
The description doesn't show clearly what is the difference between the
phyflex-imx6-1 and the phyflex-imx6-2 machine. In fact if you know the
article number, you can distinguish them.

To help a normal user the RAM configuration (one or two banks) is added
to the description, because it is recognizable on the module.

Signed-off-by: Stefan Christ <s.christ@phytec.de>