fbtest: add recipe
This is a port from ptxdist. We use only two of ptxdist patches for the
build and add a quick fix for cross building. The standard makefile from
fbtest just overrides the compile settings with false assumptions. Those
were simply removed.

This software is from pengutronix and replaces the fb-test from meta-oe.

Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>