conf: machine: include: first fix for build of phytec-qt5demo-image
The image phytec-qt5demo-image includes the package group
packagegroup-base. Through some indirections it contains the package
"firmware-imx-vpu-imx6q" and "fsl-alsa-plugins" for an imx6 build.
Since these packages are not provided in your BSP yet, the build failed.

The problem wasn't spotted earlier, because the packagegroup-base isn't
use in the phytec-hwbringup-image.

Signed-off-by: Stefan Christ <s.christ@phytec.de>