recipes-kernel: linux-mainline: remove defconfig
Remove the defconfig from the recipe. The buildprocess will use the
kernels defconfig which is given by the KERNEL_DEFAULT_DEFCONFIG
definied in the machine configuration.

Since INPUT_EVBUG isn't disabled in kernel version 3.17.6-phy2 yet, we
use a config fragment to disable it. The recipes generate the same
kernel configuration as before the patch.

Signed-off-by: Stefan Christ <s.christ@phytec.de>