recipes-graphics: gpu-viv-bin-mx6q: add Freescale demo programs
Add the Freescale demo programs in /opt/viv_samples/ as the package
'gpu-viv-bin-mx6q' to the recipe. It's not included in the root image
by default because it's roughly 600MB. Use an IMAGE_INSTALL_append in
your local.conf.

The demo programs need some additional dependences: libclc-mx6,
libvdk-mx6, libgles-mx6 and libopencl-mx6. They are added too.

Signed-off-by: Stefan Christ <s.christ@phytec.de>