scripts: use script's location to find BSP root directory
Implement search for root directory of the Yocto BSP. It uses the folder
".repo" as the anchor.

The BUILDIR is passed as a parameter to oe-init-build-env explicitly,
because the script will otherwise use the current working directory of
the process.

Now the init script should be independent of the caller's current
working directory.

Signed-off-by: Stefan Christ <s.christ@phytec.de>