classes: rename device tree file to oftree on sdcard
The device tree file on the sdcard image was originally named the same
as in the kernel tree. But this leads to an overall complex environment
in the barebox recipe folder, because the file 'env/boot/mmc' is
different for each machine, since it contains the device tree name on
the sdcard.

Therefore the device tree file is generally named 'oftree' on the sdcard
for all machines.

Signed-off-by: Stefan Christ <s.christ@phytec.de>