scripts: implement hide feature in switch_machine.py
Add the hide feature in switch_machine.py. Machines can be marked with
the tag '[hide]' in DESCRIPTION in the machine configuration. So these
machines are normally not shown in the script switch_machine.py, only if
the user uses the argument '--all'.

Also add some documentation about this feature.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>