classes: put multiple device trees on sdcard
This patch fixes the commit

  commit bc17f1c0e4135c80588a432b57bda8ee9596ff24
  Author: Stefan Christ <s.christ@phytec.de>
  Date:   Tue Apr 14 16:11:21 2015 +0200

      classes: rename device tree file to oftree on sdcard

because it accidentally removes the feature to place multiple device
tree files onto the sdcard.

Now all device trees in KERNEL_DEVICETREE are copied onto the sdcard
again and the first device tree is written to 'oftree' which is used as
the default device tree by the bootloader.

Signed-off-by: Stefan Christ <s.christ@phytec.de>