barebox/linux: add variable INC_PR
The variable INC_PR is the 'package revision' for include files. The
recipe which uses the include file must define his package revision as

    PR = "${INC_PR}.0"

Every time the include file is changed, the variable INC_PR should be
increased by the developer. That leads to a rebuild of every recipe
which uses the include file, because the recipe's PR changes
accordingly.

For more information see

   http://www.yoctoproject.org/docs/1.7/mega-manual/mega-manual.html#manually-bumping-pr

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>