barebox.inc: refactor environment directory handling
The default environment folders are added to DEFAULT_ENVIRONMENT_PATH in
.config before the script 'far' is installed into '.environment/bin'.
If the .environment folder isn't used like in the current i.MX6 build,
the '.environment' folder isn't picked by the build process and the
script 'far' isn't installed into the default environment.

The simple fix would be to move the installation of the script 'far' at
the top of the task do_prepare_env. Instead the patch refactors the
.config handling a bit.

Now only the environment is assembled in the task do_prepare_env and all
code which touches .config is done in the task do_prepare_configure.
Furthermore the inter task dependences are changed: The task
do_prepare_env is executed before the task do_prepare_configure.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>