linux: build mtd-tests kernel modules
Build mtd tests kernel modules, but don't install these modules by
default on the rootfs. That is accomplish by appending the kernel module
packages to the variable KERNEL_MODULES_RDEPENDS_BLACKLIST.

The variable MTD_TEST_PACKAGES is defined in the layer.conf and not in
the file linux.inc, because this way it can be used in the image recipe
in the layer meta-yogurt, too.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>