classes: add bbclass kconfig-prepare-configure
Merge the task do_prepare_configure, which is defined in linux.inc and
barebox.inc, into a common bbclass. There were only some minor merge
problems, e.g. a call to bbnote had to be changed, because it contains
the name 'barebox'.

The change was tested by comparing the output of 'bitbake -e' with and
without the patch. There are only date, metadata and whitespace
differences.

The variable $INC_PR of barebox.inc and linux.inc was not increased,
because the output of the recipes will not change.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>