conf: use special string ##OEROOT## in bblayers.conf.sample
The special constant '##OEROOT##' is replaced with the absolute path to
the poky root directory while the bblayers.conf.sample is installed into
build/conf/. See <poky>/meta-yocto/conf/bblayers.conf.sample for an
example.

This allows to put the build directory at any location outside the
predefined BSP folder hierarchy. The only requirement is that the meta
layers (meta-phytec, meta-qt, ...) are placed in the same folder as the
project poky.

Signed-off-by: Stefan Christ <s.christ@phytec.de>