libgles-omap3: change sgx init service
We have seen random initialization problems of the TI binary blob
drivers for the sgx core. There seems to be no archivable goal to try
debugging those binaries. The init server provided with TI's yocto
layers differs from the one provided with the graphic sdk. We switch to
the one provided with the graphic sdk and hope the issues will not come
up any more.

Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>