common: deploy config instead of defconfig
Instead of the defconfig, which is a reduce set of config symbols,
deploy the actual build configuration in the file '.config'. So a user
can easily check whether a config symbol (e.g. a driver) was enabled or
not without looking at the default value of the symbol in the specific
kconfig source code.

The patch also adds the deploying of the IPL/MLO configuration which was not
done before.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>