scripts: remove add_distro_feature in phylib
The function add_distro_feature of class BoardSupportPackage was used
for the mtd-test kernel modules and the phytec demo autostart. Since
both were replaced with different configuration approaches embedded into
the BSP the feature isn't used anymore.

Furthermore we came to the conclusion that using DISTRO_FEATURES for
dynamic build configuration is a bad thing, because it may breaks the
build in many ways.

You have to update the phyTOOLS repository, because the subclass
BSP_Builder, which inherits BoardSupportPackage, uses the member
function add_distro_feature, too.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>