fakeroot: fix standalone build of meta-phytec+soc-layers
The ti-unpack handlers needs a 32bit version of fakeroot to unpack the
ti binary archives. Therefore it depends on the recipe

    recipes-core/fakeroot/fakeroot-native32bit_1.18.4.bb

Which itself includes 'fakeroot-native_1.18.4.bb' from the
meta-openembedded layer. Since the layer is not available when
meta-phytec is build only with poky, the build fails.

Therefore the patch copies the fakeroot recipes from meta-openembedded
into this layers. If the meta-openembedded layer is available the file
search routine should select the recipes in our layer because it has a
higher priority. If only one copy of the file is available it selects
this one, too.

Signed-off-by: Stefan Christ <s.christ@phytec.de>