scripts: simplify BSP_VERSION handling
Now the Phytec's BSP_VERSION is defined in layer configuration of the
meta-phy<soc> Layers. E.g. meta-phyimx6/conf/layer.conf contains

    BSP_VERSION_imx6 = "i.MX6-PD15.1.0"

Formerly the BSP_VERSION was assembled in the script create_manifest.py
from the parameters "-r" (release) and "-p" (platform). Then it was
written to the manifest file in the xml tag "phytec" in the attribute
'release_uid'. The manifest file is hosted in the specific soc branch in
the phy2oco Repository. On checkout of the BSP with phyLinux the script
tools/init is called. Here the script write_bsp_version_to_lconf.py
parsed the manifest file in .repo/manifest.xml and extracted the
attribute 'release_uid'. Then it injected the value into the local.conf
as BSP_VERSION.

KSP projects, which are based on BSP releases, use

    BSP_VERSION_append = "-KSPXXXX-yyyy-mm-dd"

in their layer configuration instead of defining BSP_VERSION in their
local.conf.sample.

Signed-off-by: Stefan Christ <s.christ@phytec.de>