conf: machine: [imx6]: fix build of core-image-sato
This patch avoids pulling in the recipe gpu-viv-bin-mx6q for the X11
enabled build of core-image-sato in poky. It does so by removing
'gstreamer1.0-plugins-bad-meta' in the RECOMMENDS of gst-player-bin.

With the patch the image 'core-image-sato' in the distro poky builds
successfully for the i.MX6 machines.  The video file
big_buck_bunny_480p_stereo.ogg which contains a video in the theora
codec is played nicely with the gst-player in a X11 window.

Signed-off-by: Stefan Christ <s.christ@phytec.de>