conf: machine: [imx6]: remove broken udev-extraconf
The package udev-extraconf from the poky recipe udev-extraconf_1.1.bb
implements the traditional flawed approach of an automount script. It
installs the following udev rules and scripts:

    /etc/udev/rules.d/[autonet.rules,automount.rules,localextra.rules]
    /etc/udev/scripts/[network.sh,mount.sh]
    /etc/udev/mount.blacklist

The rules in automount.rules uses a 'RUN +=' statement which executes
the script 'mount.sh' which then executes the 'mount' command. This
approach is flawed. The udev 'RUN +=' statement cannot be used for
complex scripts and commands. See

    http://www.freedesktop.org/software/systemd/man/udev.html#RUN{type}
    "This can only be used for very short-running foreground tasks.
    Running an event process for a long period of time may block all
    further events for this or a dependent device.

    Starting daemons or other long-running processes is not appropriate
    for udev; the forked processes, detached or not, will be
    unconditionally killed after the event handling has finished."

The current behaviour, if the udev rules and scripts are active, is that
on sd-card insert the script executes the 'mount' command and the kernel
actually prints the correct debug messages of a successful mount, for
example

    EXT4-fs (mmcblk1p2): recovery complete
    EXT4-fs (mmcblk1p2): mounted filesystem with ordered data mode.

but the mount is not fully finished. The output of mount doesn't show
the mounted partition and it's not available in the filesystem. Maybe
the mount process is killed inbetween.

The better approach is to handle the mount operation in systemd service
unit and use the environment variable 'SYSTEMD_WANTS' in the udev rule
to start the systemd mount service on insert of the sd-card. An example
for the udev rule:

    … TAG+="systemd", ENV{SYSTEMD_WANTS}="mountdyn@%k.service"

The idea is described here

    https://bbs.archlinux.org/viewtopic.php?id=149419

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>