use := in bb files for FILESEXTRAPATHS
This change should avoid potential copy and paste errors, when a
developer copies the line

   FILESEXTRAPATHS_prepend = "${THISDIR}/${PN}/features:"

from a recipe into a bbappend.

Since '=' is used, the variable variable ${THISDIR} is expanded to the
folder of the recipe and not of the bbappend. Mostly a developer wants
to include the local folder in a bbappend.

Therefore replace all instances of '=' with ':=' (Immediate variable
expansion), so the variable ${THISDIR} is expanded to the bbappend's
folder. It doesn't harm in the bb file.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>