scripts: implement soc distinction in switch_machine.py
After executing phyLinux or while calling scripts/switch_machine.py
directly, only the machines of the selected SoC should be shown. Use the
soc information in the release manifest for that.

If the parameter --all is passed to the script switch_machine.py the
machines of all socs are shown.

    scripts/switch_machine.py  --all

Signed-off-by: Stefan Christ <s.christ@phytec.de>