conf: machine: [am335x]: fix build of core-image-sato
This patch avoids pulling in the recipe libgles-omap3 for the X11
enabled build of core-image-sato in poky. It does so by removing
'gstreamer1.0-plugins-bad-meta' in the RECOMMENDS of gst-player-bin.

With the patch the image 'core-image-sato' in the distro poky builds
successfully for the AM335x machines.  The video file
big_buck_bunny_480p_stereo.ogg which contains a video in the theora
codec is played, but slowly, with the gst-player in a X11 window.

The same fix was already applied for i.MX6 machines.

Signed-off-by: Stefan Christ <s.christ@phytec.de>