barebox: use empty directory trick for second environment folder
The barebox version for the AM335x boards uses a special trick for the
second environment overlay folder 'environment'. That approach is
better, because you cannot forget to define the needed SRC_URI_append
like

   SRC_URI_append_phyboard-mira-imx6-3 = " file://environment"

in the barebox recipe. So use this trick also for the i.MX6 barebox
version.

This trick works, because bitbake selects the most specific option for
the source file 'environment'. So the folder 'environment' in the
machine specific folder is preferred over the folder 'environment' in
the 'mx6' directory.

Acked-by: Christian Hemp <c.hemp@phytec.de>
Signed-off-by: Stefan Christ <s.christ@phytec.de>