i.MX6: machine: Add SOC_FAMILY append
Introduce the same board specific SOC_FAMILY as for the AM335x machines.
The i.MX6 SOC_FAMILYs are

   phyboard-alcor-imx6,
   phyboard-mira-imx6,
   phyboard-subra-imx6,
   phycard-imx6 and
   phyflex-imx6

Now every i.MX6 machine is in the SOC_FAMILY 'mx6' and in the SOC_FAMILY
'mx6q' or 'mx6dl' and in one of the board specific SOC_FAMILYs from
above.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>