am335x: don't put ubifs onto emmc image
Putting the ubifs into the first partition makes only sense for the
sdcard image, not for the emmc image. You don't want to flash an ubifs
onto a NAND chip when your system has booted from EMMC, because the EMMC
mostly replaces the NAND on a system.

Therefore the variable FATPAYLOAD_IMG is split into

    FATPAYLOAD_IMG_sdcard and
    FATPAYLOAD_IMG_emmc.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>