barebox.inc: remove BAREBOX_DEFAULT_ENV handling
The machine specific variable BAREBOX_DEFAULT_ENV contains the path to
the default environment in the barebox source tree. It's only used when
no other environments machineenv, boardenv and commonenv is used. It's
only a fallback mechanism.

In the first case, when machineenv, boardenv or commonenv is defined,
the variable BAREBOX_DEFAULT_ENV is not used and no files are added to
the environment from the internal barebox default environment.

In the second case, when machineenv, boardenv and commonenv are not
defined, the variable BAREBOX_DEFAULT_ENV _is_ used and the path to the
internal barebox environment is added to the config parameter

    CONFIG_DEFAULT_ENVIRONMENT_PATH

instead of the machineenv, boardenv or commonenv paths.

But the internal barebox environment in source tree in folders

    arch/arm/boards/phytec-som-am335x/defaultenv-physom-am335x/ and
    arch/arm/boards/phytec-phyflex-imx6/defaultenv-phyflex-imx6/

is nevertheless used, because our device init functions in board.c have
the code lines like

    defaultenv_append_directory(defaultenv_physom_am335x);

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>