ti33x: alsa-state: Update mixer settings
The different soundcards are identified by their names assigned in the
dts. All settings are stored in a single file and can be updated in
place. To update the mixer settings run

 $ alsactl store

and copy the asound.state file back to meta-phytec. Settings for new
soundcards can also be added in this way. We create one file per SoC so
that the file will not grow to big. This is not a technical requirement
and the files can easily be merged.

Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>