linux-common.inc: Add support for dts and dtsi in SRC_URI
This is a merge of:
[RFC v3] linux-dts.inc: use dts from src_uri

With this change it is possible to store device tree files in the
meta layers. This makes it easy to separate some project specific
hardware changes to the corresponding project layer without having to
fork the kernel.
If the change has no side effect we prepare a patch for oe-core.

Author: Jan Remmet <j.remmet@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>