classes: image_types_fsl: remove legacy code
Remove the kernel_bin_for_dtb magic which is not needed.

This patch also removes the legacy handling of device tree files in
variable KERNEL_DEVICETREE. That's the line

    DTS_BASE_NAME=`basename ${DTS_FILE} | awk -F "." '{print $1}'`

You also find this in linux-dtb.inc in poky.

Formerly you provided the full path to the dts file in the kernel source
tree like

   KERNEL_DEVICETREE = "arch/arm/boot/dts/imx6q-phytec-pbab01.dts"

Now you only provide the final name of the dtb file like

   KERNEL_DEVICETREE = "imx6q-phytec-pbab01.dtb"

Because your machines don't relay on the legacy behavior we can remove that
feature in the image creation class.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>