barebox: rewrite savedefconfig task in python
Using a python implementation gives the same workflow as for the task
'diffconfig'.  The user can simply copy&paste the path and move the
generated defconfig into his/her layer.

The patch adds a new python function py_oe_runmake for convenience. It's
a straightforward reimplementation of the bash function oe_runmake.
Since it's to hard to directly call make with an argument list (You
would have to parse ${EXTRA_OEMAKE} into an argument list), the function
just concats everything into a long string and passes it to the shell
for execution.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>