linux: Remove yocto-kernel stuff
And here ends our little excursion into the yocto-kernel world. We
experienced just too many bugs: non-generic build dependencies, config
handling, git and archiver handling, devtool and externalsrc. The yocto
kernel is very invasive in the build infrastructure and needs special
care in all the components. That is fundamental design problem. Now we
come back to the Unix philosophy of building small and modular features.

The DTS handling function had to be rewritten in python.

Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>