fsl-bin-unpack: rework Freescale's EULA handling
Instead of a self made EULA handling use new Yocto interface with
variables LICENSE_FLAGS and LICENSE_FLAGS_WHITELIST. See

    http://www.yoctoproject.org/docs/1.8/ref-manual/ref-manual.html#enabling-commercially-licensed-recipes

for details. For the user not much will change. Instead of putting

    ACCEPT_FSL_EULA = "1"

into the local.conf, now he/she has to add

    LICENSE_FLAGS_WHITELIST += "license-freescale-v4-june-2013"

to accept a single version or

    LICENSE_FLAGS_WHITELIST += "license-freescale"

to accept any version.

For now the license must be provided in an extra file 'COPYING'. In the
next gpu library version this won't be needed anymore.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>