license: Add freescale-v6-february-2015 license to license directory
Providing non generic licenses on package basis in fido does not work as
expected. To be on the save side, we copy the license to the meta-phytec
license directory. It is then available as a generic provider and will
be correctly installed for a license audit.  This gets rid of a warning
thrown by a sanity check from the license.bbclass. The code got
refactored for jethro and this licensing workaround will then not be
required.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>