conf: add i.MX6 license handling in local.conf
For the i.MX6 VPU firmware and GPU library the user has to accept the
Freescale EULA. Make this process easy by providing ready to use
variables in the local.conf.

For now we are using two different versions of the EULA for VPU and GPU.
The next GPU library version will use the same version as the VPU
firmware.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>