recipes: fix LICENSE_FLAGS matching for Freescale EULA
It should be possible to specify a wildcard for all Freescale EULA
versions in the local.conf. Therefore you have to put a underscore
between the license name and the version in the recipe.  For example:

    LICENSE_FLAGS = "license-freescale_v6-february-2015"

Then you can either accept any version of the license with

    LICENSE_FLAGS_WHITELIST += "license-freescale"

or allow only a specific version of the license with

    LICENSE_FLAGS_WHITELIST += "license-freescale_v6-february-2015"

or allow only a specific version for a specific recipe with

   LICENSE_FLAGS_WHITELIST += "license-freescale_v6-february-2015_firmware-imx"

in your local.conf. See

   http://www.yoctoproject.org/docs/1.8/mega-manual/mega-manual.html#license-flag-matching

for more details.

Signed-off-by: Stefan Christ <s.christ@phytec.de>