barebox: env: use nv variable 'rootfs' and set to 'ro'
Use a global nv variable for the generic root filesystem kernel
parameters. Some of them are common for all boot devices.

Switch "rw" to "ro". Now the kernel mounts the root filesystem readonly.
The init system will remount it rewrite based on the settings in fstab
if needed. It's important that the filesystem is readonly first because
fsck (e.g. for the ext4 filesystem) in the boot process can only run if
the filesystem is mounted readonly beforehand.

The option 'rootwait' is now used for all boot devices and not only for
"mmc". This option should not harm the other boot options since it only
delays the boot process if the rootfs cannot be found immediately.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>