barebox: env: force filesystem repair by fsck
Using the kernel parameter

    fsck.repair=yes

enables the strongest fixing mode of fsck. Documentation of fsck.repair:

  "yes" will answer yes to all questions by fsck
  (https://www.freedesktop.org/software/systemd/man/systemd-fsck@.service.html)

In fact the kernel parameter advices the systemd-fsck program to pass
the parameter '-y' to fsck.ext4. This fixes filesystem errors more
aggressively than the default '-p' (Automatically repair ("preen") the
file system).

This option should make the system more powercut and failure resistant.
Since for embedded devices there is no system administrator who can
handle a broken filesystem, the system should just try as hard as it can
to boot and start the main application. The main application can fail,
because of the broken filesystem, too, but it's better to start it than
giving up before. So it has a chance to detect errors and show an
appropriate message to the end user.

The question, what to do if the ext4 filesystem is corrupted, is open
for debate and maybe depends on the end user application, e.g.
consumer, automotive or safety critical applications.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>