conf: add note about license handling in conf-notes.txt
The i.MX6 GPU and VPU recipes now uses the Yocto based license handling
via LICENSE_FLAGS_WHITELIST. The drawback is that the error message of
bitbake, if the license is not accepted, is somehow uninformative.  The
error message doesn't directly tell the user in the last line that, he
has to accept the license via LICENSE_FLAGS_WHITELIST in the local.conf.

    ERROR: Nothing PROVIDES 'virtual/libgles2'
        (but .../meta-qt5/recipes-qt/qt5/qtbase_git.bb DEPENDS on or otherwise requires it)
    ERROR: imx-gpu-viv PROVIDES virtual/libgles2 but was skipped:
        because it has a restricted license not whitelisted in LICENSE_FLAGS_WHITELIST
    ERROR: libgles-omap3 PROVIDES virtual/libgles2 but was skipped:
        incompatible with machine phyflex-imx6-2 (not in COMPATIBLE_MACHINE)
    NOTE: Runtime target 'qtbase' is unbuildable, removing...
    Missing or unbuildable dependency chain was: ['qtbase', 'virtual/libgles2']
    ERROR: Required build target 'phytec-qt5demo-image' has no buildable providers.
    Missing or unbuildable dependency chain was:
        ['phytec-qt5demo-image', 'qtbase', 'virtual/libgles2']

    Summary: There were 2 ERROR messages shown, returning a non-zero exit code.

Therefore rework the conf-notes.txt file, which is shown everytime
before sourcing the build environment. Add some text that the user
should have to look into the local.conf file. Hopefully this catches
some user support requests.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>