barebox: env: use 'data=journal' for ext4 filesystem for reliability
Change ext4 data mode from 'ordered' to 'journal', because we observe
that newly written files are empty, after a powercut happens.  This
change improves the filesystem reliability in case of a power cut.

There are three modes for ext4 filesystem:

    see https://www.kernel.org/doc/Documentation/filesystems/ext4.txt

    * writeback mode (journal is not used)                    (fastest)
    * ordered mode   (only metadata is committed to journal)
    * journal mode   (data+metadata are committed to journal) (slowest)

The first observation of reliability issue were empty ssh host keys in
/etc/ssh.  These are generated on the first ssh login via ssh-keygen. If a
powercut happens right after the keys are created, the files are empty on the
next boot.  That leads to ssh login failures, because keys are invalid and
ssh-keygen doesn't regenerated key files that already exist.

This option has be set via rootflags in the kernel commandline, because
it must be present at the initial mount of the filesystem. It cannot be
set in the fstab, because changing this mode via remounting in service
systemd-remount-fs.service fails.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>