phygittag: fix BRANCH for upstream tags
It should be possible to use upstream tags in your barebox/kernel
recipes for our integration branches. A recipe with uses the upstream
tag directly without phytec local patches may look like:

    # recipe filename linux-mainline_4.1.18.bb
    inherit phygittag

    GIT_URL = "git://git.phytec.de/${PN}"
    SRC_URI = "${GIT_URL};branch=${BRANCH};protocol=git"
    ...

The phygittag bbclass will generate

    BRANCH = "v4.1.18-phy"
    GIT_TAG = "v4.1.18"

This setup assumes that the tag "v4.1.18" is available in the
repository

    git://git.phytec.de/linux-mainline

and the branch "v4.1.18-phy" exists in this repository.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>