classes: add patch-auto-append.bbclass
In a KSP project you have to mostly patch the kernel. If there is no
extra kernel repository, the patches are included in the meta layer.
While development it's annoying to constantly update the list of patches
in the SRC_URI. Especially if the name of a patch file changes
afterwards.  Here the patch-auto-append class comes to rescue.
Just include the three lines

     inherit patch-auto-append
     PATCH_AUTO_APPEND_DIRS_prepend := "${THISDIR}/${PN}"
     FILESEXTRAPATHS_prepend := "${THISDIR}/${PN}:"

in your bbappend and you never have to update the SRC_URI again.

Signed-off-by: Stefan Christ <s.christ@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>